More inspection.

+++ main/cpp/demo-operator_overloading/functor-info/distance.cpp
@@ -2,4 +2,4 @@
 
 #include <cmath>
 
-int Distance::operator()(int x, int y) const { return sqrt(pow(fromX - x, 2) - pow(fromY - y, 2)); }
+int Distance::operator()(int x, int y) const { return static_cast<int>(sqrt(pow(fromX - x, 2) - pow(fromY - y, 2))); }
